[#395] Move segment tabs wrapper into a layout component
