+++ frontend/src/pages/cases/layout/index.js
@@ -1,1 +1,2 @@
 export { default as CaseWrapper } from "./CaseWrapper";
+export { default as SegmentTabsWrapper } from "./SegmentTabsWrapper";
